requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pillow==11.1.0
pillow-avif-plugin==1.5.1
python-multipart==0.0.20
websockets==15.0.1
